Fixed type mismatch

+++ include/mp3.h
@@ -8,6 +8,6 @@ typedef uint8_t byte;
 
 int read_header(uint64_t position, const char* filename);
 
-uint32_t calculate_frame_length(uint16_t bit_rate, uint16_t sample_rate, byte padding);
+uint32_t calculate_frame_length(uint32_t bit_rate, uint32_t sample_rate, byte padding);
 
 #endif // MP3_H
